refactor: Test 코드의 패지키 구조를 변경합니다
